Organize test code for run.ui package.
