+++ skaffold/src/test/kotlin/com/google/container/tools/skaffold/run/ui/BaseSkaffoldSettingsEditorTest.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.container.tools.skaffold.run
+package com.google.container.tools.skaffold.run.ui
 
 import com.google.common.truth.Truth.assertThat
 import com.google.container.tools.skaffold.SkaffoldFileService
-import com.google.container.tools.skaffold.run.ui.BaseSkaffoldSettingsEditor
+import com.google.container.tools.skaffold.run.AbstractSkaffoldRunConfiguration
 import com.google.container.tools.test.ContainerToolsRule
 import com.google.container.tools.test.TestService
 import com.google.container.tools.test.UiTest
